Periodically run the map reduce tasks to update summaries, using task queue.
This is not the best way to do it. A cron job would be both simpler and more stable, but I just wanted to try the task queue feature. Still has several issues, for instance it will break when the task accidentally get run twice because this will result in two chains of tasks that each schedule a new task when they run. Instead tasks should somehow be made idempotent.